+++ inc/Helpers/DisplayImplementation.php
@@ -35,6 +35,70 @@ class DisplayImplementation {
 	}
 
 	/**
+	 * Renders the attributes for an HTML tag.
+	 * The function accepts attributes in associative array,
+	 * filters out falsy attributes and escapes the attribute values.
+	 *
+	 * @param array $attributes List of attributes to render.
+	 * @return string
+	 */
+	public function attributes( array $attributes = [] ): string {
+		$attributes = array_map(
+			function ( $name, $value ): string {
+				// All the falsy attributes should be filtered out except the alt attribute for the images.
+				if ( ! boolval( $value ) && ! in_array( $name, [ 'alt' ], true ) ) {
+					return false;
+				}
+
+				if ( true === $value ) {
+					return $name;
+				}
+
+				if ( is_callable( $value ) ) {
+					// phpcs:ignore NeutronStandard.Functions.VariableFunctions.VariableFunction
+					$value = $value();
+				}
+
+				if ( is_array( $value ) || is_object( $value ) ) {
+					$value = wp_json_encode( $value );
+				}
+
+				return $name . '="' . esc_attr( $value ) . '"';
+			},
+			array_keys( $attributes ),
+			array_values( $attributes )
+		);
+
+		return join( ' ', array_filter( $attributes ) );
+	}
+
+	/**
+	 * Renders data attributes for an HTML tag.
+	 *
+	 * @param array $attributes List of data attributes to render.
+	 * @return string
+	 */
+	public function data_attributes( array $attributes = [] ): string {
+		$attributes = array_map(
+			function ( $name, $value ): string {
+				if ( true === $value ) {
+					$value = 'true';
+				} elseif ( false === $value ) {
+					$value = 'false';
+				} elseif ( is_array( $value ) || is_object( $value ) ) {
+					$value = wp_json_encode( $value );
+				}
+
+				return 'data-' . $name . '="' . esc_attr( $value ) . '"';
+			},
+			array_keys( $attributes ),
+			array_values( $attributes )
+		);
+
+		return join( ' ', array_filter( $attributes ) );
+	}
+
+	/**
 	 * Render a reusable block post.
 	 *
 	 * @param int|WP_Post|null $post The reusable block post.