+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.5.0] - 2023-07-14
+
 ### Added
 
 - Added Helpers\Display::attributes and Helpers\Display::data_attributes - render attributes for HTML elements
@@ -57,7 +59,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added Options utilities
 - Added Features classes
 
-[unreleased]: https://github.com/airfleet/airfleet-wordpress-framework-php/compare/0.4.0...main
+[unreleased]: https://github.com/airfleet/airfleet-wordpress-framework-php/compare/0.5.0...main
 [0.1.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/0.1.0
 
 [0.2.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/0.2.0
@@ -67,3 +69,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [0.3.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/0.3.0
 
 [0.4.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/0.4.0
+
+[0.5.0]: https://github.com/airfleet/airfleet-wordpress-framework-php/releases/tag/0.5.0